chore: do not require IAMMAnager permission for terraform
Unfortunately, creating an application and the associated access key
requires the IAMManager permission, which is an organization-level
right that gives full access to the security of the organization.

So instead of automatically creating a Scaleway application for
Airflow's object storage access, we now require extra variables
that point to a pre-created Scaleway application dedicated to
Airflow (application ID, access key and secret key).