chore: chown airflow files to group 0
According to the documentation for the Airflow Docker image
( https://airflow.apache.org/docs/docker-stack/build.html#important-notes-for-the-base-images )
files need to be owned by user `airflow` and group `root` (0).

Group `airflow` doesn't exist in the image, which apparently caused
Docker to fall back to group 0 anyway, so this change should not have
any impact.