fix: correctly create PG extensions
Because docker-entrypoint-initdb.d/10_postgis.sh was marked as
executable, it was executed instead of sourced by docker-entrypoint.sh
causing the "${psql[@]}" syntax to fail. This also caused the processing
of the init scripts to be aborted, causing 10_unaccent.sh not to be
executed at all.

Note that this was not detected until one tried to use the expected
extensions (plpython3u, unaccent), since only the first launch of
the postgres container (where the DB was created) failed.