fix(pipeline): Do not require 'frais' to be not null
The column is not required in the DORA datascheme in the database nor is
enforced to be non null in the DORA API that exports the data to us.

Moreover, the 'frais' are also not mandatory in data-inclusion-schema,
so everything points in the direction of the constraint to be lifted.

Still add a `not_constant` test to avoid the silly case of only NULL
values, for instance.