chore(root) : Fix docker-compose base environment values
docker-compose up returns "contains true, which is an invalid type, it
should be a string, number, or a null" in version 1.29.2.

The spec says :
https://github.com/compose-spec/compose-spec/blob/master/spec.md#environment

And indeed, in the map version (not the list in key=value) for the YAML
properties, the `True` value must be quoted.

I don't really understand how it worked before, but maybe was it just a
warning.