fix(tox) : Stop specifying the Python version
Recent versions of Linux all ship Python 3.11 which is out since October
2022.

Drop tox's envlist that makes it mandatory to have __that__ version
to run the tests.

There is very little chance that the few Python tests we have will be
sensitive to the version or Python, or we'd be doing some pretty nasty
code.