chore(dags) : Use pg helpers in the other DAGs.
Some of the DAGs (for instance all the INSEE ones)
could maybe be combined ?

I am a little surprised that for those ones we don't
have the datalake layer. I agree those are mostly
for "sources" in the DI sense but technically, it
would be nice to assume that we can reproduce everything
from a single S3 dump.

Maybe that's too much ? I still think I would sleep
better at night if the Extract + Load was always the same.

In which case we could separate the subfolders though.
- di-sources
- seeds
- you name it