chore(pipeline): update xp data
The xp data was previously extracted from dora in staging.
This hack has saved us some time, but it's time to stop.
First, the data could be wipped by the team at dora.
Second, it's difficult to update.
Following the migration of the xp data from dora staging to a clean
google sheet, this commit updates transformation accordingly.