chore(pipeline): use pydantic validation
Currently validation differs between the pipeline and the api.
The api relies on the schema validation, using pydantic. But
the pipeline has its own sql-based validation. This can lead
to inconsistencies between pipeline and api and duplicates code.

This commit leverages plpython to reuse the schema validation
in the pipeline.

Validation is done at the source level. This commit also materializes
validation errors in a dedicated model, rather than using dbt data
tests, which are less convenient for products to use.