chore(pipeline) : Add a wait between Brevo contact sync & notify
Make it dead simple for now.

- the notifications were not sent, even though the mailing list attached
  seemed OK in the interface
- we update the list in the same DAG as the notification dispatch
- we suppose that the list is not "properly updated" server-side at the
  time we send the notifications (race condition) so we introduce a
  fixed wait time for now.

We checked that re-triggering the notification alone (without the ML
sync) was indeed working, so chances are on our side. If the MLs are
dispatched next month properly, we could work on a mechanism that
actually checks the ML contents but... not sure we should bother.

BTW TimeDeltaSensor was interesting but unfortunately does not work
super well as the delta is computed from the start date of the DAG which
is "somewhere in 2023", so it does not wait at all in its current form
(without changing much of the DAG shape and initial params), so we took
a quick and simple (dirty ?) approach.