chore(pipeline) : Make the Dora contact source work again
During build, the Dora intermediate table is first built (like any other
source) in the `public_intermediate_tmp` schema.

This is done at the DBT project level, which means that any table from
the "intermediate" layer is then supposed to reside (as far as `ref()`
is concerned) in the `public_intermediate_tmp` schema at that moment.

Unfortunately, the Brevo table is not yet built in that schema as its
DAG is still just immediately running the models.

It feels normal to just follow the exact same logic and have a
pre-built, tested, and then ran model for the Brevo things.

By the way, we now have to remove any contact from Brevo that expressed
a will to be removed from the display lists; it has not been done yet
for mes-aides and mediation-numerique.