chore(ci) : Deploy to staging through a PR label
It is annoying to be competing with each other when pushing code, so I
feel like it's going to help if we decide to deploy through adding a
non-mandatory label than "all the time unless draft".

Also, drafts mean something different IMHO. They should be used as
"please do not review yet, this is a work in progress, open for
discussion".

This commit also splits the jobs in two:

- one job will be responsible of building and deploying images to
  staging
- another one will deploy to prod, the 'release' branch only.

Reusable actions have been used to ensure DRYness.