chore(pipeline) : Make the main DAG run hourly
Having two DAGS running the geocodages model in the same time doesn't
bode well with the incremental models.

Let's switch to running the main DAG hourly, which also kind of makes
sense as our pipeline is more frequently updated and is also by far the
simpler option, thus the one with the fewer surprises.