chore(dev) : Do not restart docker images in dev
The 'on-failure' setting makes the images restart almost all the time,
especially upon reboot, if I work on this project or not.

This is good in a production setting, but in development, restarting
after a reboot on my personal computer drains my battery for no good
reason.

Furthermore, if there was an actual issue with the containers in
development, they should  just stop and the developer should fix the
issue, not run into a mindless loop till oblivion.