fix(pipeline) : Fix soliguide,monenfant zone_diffusion issue
When the zone_diffusion_code is not filled, the services are not
returned.

Following a fix about those zones in
b0c11dddcea7569e6a2acb27249e9a0ee035b703, there was a tiny mistake (an
extra 's') that made 50_000 services become undiscoverable.

Since that can become a huge issue, let's enforce better data testing in
the pipeline to ensure it does not arise again.