chore(pipeline) : Fix unit tests
Note : the code was incorrect, thank us for unit testing it.

Side note : ARRAY is unsupported for the `dict` data type in unit tests,
currently, so we have to use SQL, meaning we also have to define each
and every field out there.

Other side note : as a side effect, this checks that the 'profils' field
is made empty when going from 'stg' to 'int' layer with DORA services.