feat(pipeline) : Simplify notify_rgpd_contacts
After an entire year of operation, we noticed that our sources never
change their contact emails. Ever.

Even if someday this happens, we can probably suppose that it's till
going to be quite rare.

In those exceptional cases, it means that the structure would get a new
RGPD notification.

But it does make the code way easier to maintain and more robust:
- now we will never miss an email that would have been "forgotten" by
  Brevo (due to rate limits, payment issues, ...)
- it can also be seen as a feature to send a new notification if the
  email changes, for instance when the previous one was incorrect.

Moreover, we don't need to maintain the `contact_uid` list that was sent
to brevo.