fix(api): fix flaky tests
These two tests 'snapshot' the API response. The response depends on objects (service structures) that are pseudo-randomly generated using factories. In general, it works well: values are generated randomly but deterministically, so when we rerun the tests, the results are reproducible!

However, as soon as we make any changes to the factory or the SQLAlchemy model, the order of value generation shifts, and the values change.

The solution: hard-code the structures/services in these two tests instead of relying on the factory.
Credit to vmttn for those explanation