chore(pipeline) : use Sentry as main notifier
Sentry is a much better tool than our little code executor, at least in
terms of not spamming our alerting channel. We can precisely configure
how and when we want to be notified, what is important, discuss about
issues, etc.

It also now directly links every issue to a particular commit, making it
easier to identify regressions.

The flow changes a little bit as the logs are in the Sentry issue, not
directly linked in the Slack message, but I hope we'll be able to triage
much better.