fix(deployment) : Actually enable request tracing
There was a SENTRY_ENABLE_TRACING parameter introduced in
65a4d4e2c095956d676e4bd70a3d3347f7d435ba, but it is now
removed from the Sentry SDK documentation.

The correct configuration is now the traces_sample_rate setting:
https://github.com/getsentry/sentry-docs/pull/12439

It's a little bit disappointing to notice though that having the traces
enabled since a long time didn't help getting the ooutgoing HTTP
requests timings, as those are not supported in airflow-sentry.