feat(api) : Precompute cluster_master_id to improve performance
With this column we can very efficiently select only the highest quality
or date_maj among each cluster_id group, voiding the need for the
previous double self-join + subqueries.

It will be very useful as well for the services.

It also means the `cluster_id` or `doublons_groupe_id` is not needed
anymore, so we remove it from the API and database.

Some marginal fixes are also included (a limit case in tests, an
improvement to StructureFactory)