fix(datawarehouse) : Fix DNS resolution through caching
The previous version was not fully working as the DNS requests were
massively failing.

After some careful research, it seems that it's not safe to use that
many threads and aggressively hit the DNS servers, especially since most
of the time we hit the same base hosts (francetravail.fr, ...) on
different paths.

We reduced the "normal" batch size to 1000 (seems ok, the current total
URLS are ~30_000)

We added a local, in-memory DNS cache to stop being so agressive with
the poor lad.