fix(pipeline): fix airflow image build
There was a glibc version mismatch between the compile and
runtime images. It looks like the airflow image uses an
outdated glibc version. The image does not get updated
as frequently as it used to, now that airflow 3 is out.

Upgrading the package manually may have been enough, but
removing the compile image has been favoured as an opportunity
to simplify things. Ideally the image customization should be
minimal. This would make it easier to deploy airflow somewhere
else.