chore(api): Attempt at fixing the settings object issue
We want to override setting on the object instance attached to the app
fixture instance, but most of the code actually directly accesses the
module's value instead, which has been set beforehand and thus is not
modified.

Either I'm changing the whole app to only use references through the
app's state (with heavy implications in the authentication code) or I
restore the monkeypatch version for this object.