fix(deployment): Separate application env vars and deployment vars
This also means that we can't use terraform logic as easily as before to
generate postgres URLs or S3 urls derived from the created bucket.

Thus I chose to slightly simplify the bucket name to not need a suffix
anymore. Why was it necessary in the first place ?

We can now simply define a secret in the env file format containing all
the application secrets at once.