chore(pipeline): fix siret succession handling
The int__sirets table was slightly incorrectly computed; in some cases
we could not complete the check for a successor because the latest
state of an etablissement was pre-filtered, leading to ~700 errors in the
thankfully introduced test in the "marts" step.

Those are now fixed and the SIRETs are rewritten to match the latest
open etablissement in all cases.