chore(pipeline): Fix model validation
The current v1 pipeline implementation has a few pitfalls, one of them
being that we share the same field names for frais & thematiques in v0 &
v1 but not the same column names in the database processing stages.

We have to take that into account beforeapplying a validation that
awaits specific column names to match schema fields.