chore(pipeline): go back to v0 files for action-logement
It was unnecessary to go through the whole v1 mapping for v1 just for
the lien_source. Go back to "old" v0 files (fixing the fact that we only
had the structure file link originally though) and add the missing
columns.

This way we don't lose all the other columns.