chore(pipeline): Fix monenfant file loading
The previous expression is a lazy evaluation, not a list.
Also, s3_hook.downlmoad_file() returns a string, not a Path;
but out helper wants a Path.

Make it clear that we download the files first (with a list and not a
generator) and then load the contents of those files in another go.