revert(pipeline): Remove OR condition in URL matching
This reverts the OR condition introduced in a45c145d9c2c68289415535f19598fa6acffc04a.

The original commit claimed to fix a case where "URLs would never be tried
again", but this was likely a mistake/misunderstanding of the initial
problem.

With the old code (no OR), if an incoming URL didn't match any input_url,
it was simply treated as a new URL and checked - not "never tried".

The only reason I can find for that commit is "I was drunk".

The added OR condition caused issues:
	1. Duplicate rows were introduced (hence the DISTINCT)
	2. No cleanup of stale rows

We'll also add a post hook to clean stale lines.